+++ contracts/tasks/index.ts
@@ -9,3 +9,4 @@ import './cancelRound'
 import './evmIncreaseTime'
 import './auditTally'
 import './fetchRound'
+import './mergeAllocations'
